+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
 - `drawer` 组件 `with-header` 配置项改为 `showHeader`
-- `drawer` 组件新增 `showCloseBtn` 和 `showFooter` 配置项
+- `drawer` 组件新增 `showCloseBtn` 配置项
 - `drawer` 组件新增 `modal-blur` 配置项
 - `dialog` 组件新增 `modal-blur` 配置项